requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
ollama==0.1.6
python-dotenv==1.0.0
